master

+++ question_3.java
@@ -4,24 +4,24 @@ import java.util.Scanner;
 public class question_3 {
 
     public static void main(String[] args) {
+        
         Scanner keyboard = new Scanner(System.in);
 
-        int n = keyboard.nextInt();
-        int k = keyboard.nextInt();
+        long n = keyboard.nextInt();
+        long k = keyboard.nextInt();
 
-
-        int[] shit = new int[n];
+        int[] shit = new int[(int)n];
         for (int i = 0; i < n; i++) {
             shit[i] = keyboard.nextInt();
         }
 
-        int smallest = k * 2;
+        long smallest = k * 2;
         ArrayList < Integer > list = new ArrayList < Integer > ();
 
         //checks every number in the array
         for (int x = 0; x < n; x++) {
 
-            int timeBetween = Math.abs((shit[x] % k) * 2 - k);
+            long timeBetween = Math.abs((shit[x] % k) * 2 - k);
 
             //checks if the number is smaller than the smallest
             if (timeBetween < smallest) {
@@ -37,8 +37,6 @@ public class question_3 {
         for (int x = 0; x < list.size(); x++) {
             System.out.print(list.get(x) + " ");
         }
-
-
     }
 
 }
